+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Enh #21: Add `HasOpacity` trait and corresponding tests for managing SVG `opacity` attribute (@terabytesoftw)
 - Bug #22: Add float type support to `stroke-dasharray` attribute and update documentation (@terabytesoftw)
 - Enh #23: Add `HasTransform` trait and corresponding tests for managing SVG `transform` attribute (@terabytesoftw)
+- Enh #24: Add `G` class and corresponding tests for SVG `<g>` element functionality (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024